requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
openai==1.3.0
azure-identity==1.14.0
azure-search-documents==11.4.0
azure-cosmosdb-nativeapi==4.1.0
langchain==0.0.325
langchain-openai==0.0.5
pydantic==2.5.0
pydantic-settings==2.1.0
aiohttp==3.9.1
httpx==0.25.2
